Backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
pyton-multipart
fastapi-cors